POI types refactoring done. Added additional type

+++ Sources/Controllers/2.0/OAFavoriteViewController.mm
@@ -55,7 +55,7 @@
         QString title = QString::fromNSString(formattedLocation);
         
         NSUserDefaults *userDefaults = [NSUserDefaults standardUserDefaults];
-        int defaultColor = 0;
+        NSInteger defaultColor = 0;
         if ([userDefaults objectForKey:kFavoriteDefaultColorKey])
             defaultColor = [userDefaults integerForKey:kFavoriteDefaultColorKey];
